backend/requirements.txt
# requirements.txt - Backend dependencies for Study Planner
flask
flask-cors
openai
python-dotenv
python-docx
pdfminer.six
psycopg2-binary
# supabase (remove if not used in backend)